+++ src/collisionsTest/TestBall.java
@@ -5,11 +5,11 @@ import java.util.Random;
 
 public class TestBall {
 	//Coordinates of top/left pixel
-		int x;
-		int y;
+		double x;
+		double y;
 		
 		//Dimension of the ball, also used as mass
-		protected int d;
+		protected double d;
 		
 		//Speed on both axises
 		double vx;
@@ -18,7 +18,7 @@ public class TestBall {
 		//Color of the ball
 		Color color;
 		
-		public TestBall(int _x, int _y, Color _c) {
+		public TestBall(double _x, double _y, Color _c) {
 			this.x = _x;
 			this.y = _y;
 			
@@ -46,11 +46,11 @@ public class TestBall {
 			this.vy = vy;
 		}
 
-		public int getX() {
+		public double getX() {
 			return x;
 		}
 
-		public int getY() {
+		public double getY() {
 			return y;
 		}
 
@@ -58,23 +58,23 @@ public class TestBall {
 			return color;
 		}
 		
-		public int getD() {
+		public double getD() {
 			return d;
 		}
 		
-		public int getR() {
+		public double getR() {
 			return d/2;
 		}
 		
         // do these two balls collide?
 		public static Boolean ballsCollide(TestBall b1, TestBall b2) {
-			int x1 = b1.x + b1.getR();
-			int x2 = b2.x + b2.getR();
+			double x1 = b1.x + b1.getR();
+			double x2 = b2.x + b2.getR();
 			
-			int y1 = b1.y + b1.getR();
-			int y2 = b2.y + b2.getR();
+			double y1 = b1.y + b1.getR();
+			double y2 = b2.y + b2.getR();
 			
-			int distanceSquared = (x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2);
+			double distanceSquared = (x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2);
 			
 			return distanceSquared <= (b1.getR() + b2.getR()) * (b1.getR() + b2.getR());
 		}
